+++ api/src/test/java/ai/djl/test/mock/MockNDArray.java
@@ -752,7 +752,7 @@ public class MockNDArray implements NDArray {
 
     /** {@inheritDoc} */
     @Override
-    public NDList split(int sections, int axis) {
+    public NDList split(long sections, int axis) {
         NDList list = new NDList();
         for (int i = 0; i < sections; i++) {
             list.add(new MockNDArray(manager, device, getShape().slice(1), dataType, sparseFormat));
@@ -762,7 +762,7 @@ public class MockNDArray implements NDArray {
 
     /** {@inheritDoc} */
     @Override
-    public NDList split(int[] indices, int axis) {
+    public NDList split(long[] indices, int axis) {
         return null;
     }
 
@@ -785,13 +785,13 @@ public class MockNDArray implements NDArray {
 
     /** {@inheritDoc} */
     @Override
-    public NDArray expandDims(long axis) {
+    public NDArray expandDims(int axis) {
         return null;
     }
 
     /** {@inheritDoc} */
     @Override
-    public NDArray squeeze(long[] axes) {
+    public NDArray squeeze(int[] axes) {
         return new MockNDArray();
     }
 
@@ -821,7 +821,7 @@ public class MockNDArray implements NDArray {
 
     /** {@inheritDoc} */
     @Override
-    public NDArray argSort(long axis, boolean ascending) {
+    public NDArray argSort(int axis, boolean ascending) {
         return null;
     }
 
@@ -839,13 +839,13 @@ public class MockNDArray implements NDArray {
 
     /** {@inheritDoc} */
     @Override
-    public NDArray softmax(long[] axes, double temperature) {
+    public NDArray softmax(int[] axes, double temperature) {
         return null;
     }
 
     /** {@inheritDoc} */
     @Override
-    public NDArray logSoftmax(long[] axes, double temperature) {
+    public NDArray logSoftmax(int[] axes, double temperature) {
         return null;
     }
 
@@ -959,7 +959,7 @@ public class MockNDArray implements NDArray {
 
     /** {@inheritDoc} */
     @Override
-    public NDArray swapAxes(long axis1, long axis2) {
+    public NDArray swapAxes(int axis1, int axis2) {
         return null;
     }
 
@@ -971,7 +971,7 @@ public class MockNDArray implements NDArray {
 
     /** {@inheritDoc} */
     @Override
-    public NDArray transpose(long... dimensions) {
+    public NDArray transpose(int... dimensions) {
         return null;
     }
