+++ integration/src/main/java/ai/djl/integration/tests/model_zoo/object_detection/SingleShotDetectionTest.java
@@ -101,7 +101,7 @@ public class SingleShotDetectionTest {
                                                     classPredictions,
                                                     boundingBoxPredictions,
                                                     anchors));
-                            return detections.singletonOrThrow().split(new int[] {1, 2}, 2);
+                            return detections.singletonOrThrow().split(new long[] {1, 2}, 2);
                         }));
         return ssdPredict;
     }
